# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SmartLib Pro - Library Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reservations:
    get:
      operationId: listReservations
      tags: [reservations]
      summary: List reservations
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, fulfilled, cancelled, expired]
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Reservations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationList"
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Reserve a book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /reservations/{id}/cancel:
    post:
      operationId: cancelReservation
      tags: [reservations]
      summary: Cancel reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"

  # Fines
  /fines:
    get:
      operationId: listFines
      tags: [fines]
      summary: List fines
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [unpaid, paid]
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Fines list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FineList"
  /fines/{id}/pay:
    post:
      operationId: payFine
      tags: [fines]
      summary: Mark fine as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fine paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fine"

  # Notifications
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, librarian, student]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        studentId:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Book:
      type: object
      required: [id, title, isbn, totalCopies, availableCopies, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        isbn:
          type: string
        description:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        totalCopies:
          type: integer
        availableCopies:
          type: integer
        publishedYear:
          type: ["integer", "null"]
        language:
          type: string
        shelfLocation:
          type: ["string", "null"]
        authorId:
          type: ["integer", "null"]
        categoryId:
          type: ["integer", "null"]
        publisherId:
          type: ["integer", "null"]
        author:
          $ref: "#/components/schemas/Author"
        category:
          $ref: "#/components/schemas/Category"
        publisher:
          $ref: "#/components/schemas/Publisher"
        createdAt:
          type: string
          format: date-time
...
    Author:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        bio:
          type: ["string", "null"]
...
    Category:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
...
    Publisher:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
...
    BorrowRecord:
      type: object
      required: [id, userId, bookId, borrowedAt, dueDate, status]
      properties:
        id:
          type: integer
        userId:
          type: integer
        bookId:
          type: integer
        borrowedAt:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
        returnedAt:
          type: ["string", "null"]
          format: date-time
        renewCount:
          type: integer
        status:
          type: string
          enum: [active, returned, overdue]
        user:
          $ref: "#/components/schemas/User"
        book:
          $ref: "#/components/schemas/Book"
...
    Reservation:
      type: object
      required: [id, userId, bookId, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        bookId:
          type: integer
        status:
          type: string
          enum: [pending, fulfilled, cancelled, expired]
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: ["string", "null"]
          format: date-time
        user:
          $ref: "#/components/schemas/User"
        book:
          $ref: "#/components/schemas/Book"
    ReservationInput:
      type: object
      required: [userId, bookId]
      properties:
        userId:
          type: integer
        bookId:
          type: integer
    ReservationList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Reservation"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Fine
    Fine:
      type: object
      required: [id, userId, borrowRecordId, amount, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        borrowRecordId:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [unpaid, paid]
        paidAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
        borrowRecord:
          $ref: "#/components/schemas/BorrowRecord"
    FineList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Fine"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Notification
    Notification:
      type: object
      required: [id, userId, title, message, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [due_reminder, overdue_alert, reservation_ready, fine_issued, general]
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # Activity
...